[Rust][refactor](error): Removed error module.
