+++ bindings/rust/wasmedge-sys/src/instance/memory.rs
@@ -6,8 +6,10 @@
 //! restricts the size to which the memory can grow later.
 
 use crate::{
-    error::{check, MemError, WasmEdgeError},
-    ffi, WasmEdgeResult,
+    error::{MemError, WasmEdgeError},
+    ffi,
+    utils::check,
+    WasmEdgeResult,
 };
 use std::ops::RangeInclusive;
 
@@ -112,7 +114,8 @@ impl Memory {
     /// then an error is returned.
     ///
     /// ```
-    /// use wasmedge_sys::{error::{CoreError, CoreExecutionError, WasmEdgeError}, Memory, MemType};
+    /// use wasmedge_sys::{Memory, MemType};
+    /// use wasmedge_types::error::{CoreError, CoreExecutionError, WasmEdgeError};
     ///
     /// // create a Memory: the min size 1 and the max size 2
     /// let ty = MemType::create(1..=2).expect("fail to create a memory type");
